refactor: move Dialog query for available Content Scopes into Dialog Content, rename content component name and move Dialog into parent
